Fix long error with long to int cast

+++ Hugger Application/Hugger Application/Data/Repository/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace Hugger_Application.Models.Repository
     public interface IUserRepository : IRepository<User>
     {
         Task<User[]> GetAllUsersAsync(bool includeMatches=false);
-        Task<User> GetUserByIDAsync(int userID, bool includeMatches = false);
+        Task<User> GetUserByIDAsync(long userID, bool includeMatches = false);
 
         Task<User[]> GetAllUsersByLocalizationIDAsync(bool includeLocalization = false);
         Task<User> GetUserByLoginAsync(string login);
